Use isinstance(x, t) instead of type(x)==t
These checks were broken anyway, because they only checked for
float, while the values where actually numpy.float64. As a result,
many functions returned an an array with a single element. This
array then got silently converted to a float anyway, but this will
change, as this warning is emitted:

DeprecationWarning: Conversion of an array with ndim > 0 to a scalar
is deprecated, and will error in future. Ensure you extract a single
element from your array before performing this operation.
(Deprecated NumPy 1.25.)

isintance will return True for both float and numpy.float64, so
now everything will work as intended.

Also removing a mutable array from a function definition.