Make ScopeSim_Templates test work without internet with -m "not webtest"
Significant changes to cluster_utils, but that is not used outside of
ScopeSim_Templates, so that should be fine.

A bit refactoring in test_validate_content because paramaterized
fixtures do not exist in pytest.

And a lot of webtest markers.

Verified that no internet is required in a guix container without internet.